[WIP] Working at new login screen.
